+++ cli/cmdssh/sshexisting.go
@@ -8,7 +8,6 @@ package cmdssh
 import (
 	"fmt"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/alimtvnetwork/gitmap-v28/cli/apperror"