+++ cli/cmdssh/sshjoin_rm_cmd.go
@@ -3,7 +3,6 @@ package cmdssh
 import (
 	"context"
 	"fmt"
-	"strings"
 
 	"github.com/alimtvnetwork/gitmap-v28/cli/apperror"
 	"github.com/spf13/cobra"